+++ mylibft/libft/ft_types.h
@@ -6,15 +6,15 @@
 /*   By: andeviei <andeviei@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/24 12:48:44 by andeviei          #+#    #+#             */
-/*   Updated: 2023/11/24 19:12:44 by andeviei         ###   ########.fr       */
+/*   Updated: 2023/11/24 23:06:56 by andeviei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_TYPES_H
 # define FT_TYPES_H
 
-# define FT_INT_MAX ((int)2147483647)
-# define FT_INT_MIN ((int)-2147483648)
+# define FT_INT_MAX 2147483647
+# define FT_INT_MIN -2147483648
 
 typedef enum e_bool {
 	FALSE = 0,
